Destroy all tax adjustments in the order adjuster
As a mentioned in https://github.com/solidusio/solidus/issues/909,
the current implementation of the taxation system destroys
tax adjustments one item at a time.

This commit changes that so all the order is stripped of
all tax adjustments, resulting in less database calls.

+++ core/app/models/spree/tax/item_adjuster.rb
@@ -15,6 +15,7 @@ module Spree
         @rates_for_order_zone = options[:rates_for_order_zone]
         @rates_for_default_zone = options[:rates_for_default_zone]
         @order_tax_zone = options[:order_tax_zone]
+        @skip_destroy_adjustments = options[:skip_destroy_adjustments]
       end
 
       # Deletes all existing tax adjustments and creates new adjustments for all
@@ -27,7 +28,7 @@ module Spree
       def adjust!
         return unless order_tax_zone(order)
         # Using .destroy_all to make sure callbacks fire
-        item.adjustments.tax.destroy_all
+        item.adjustments.tax.destroy_all unless @skip_destroy_adjustments
 
         TaxRate.store_pre_tax_amount(item, rates_for_item)
